requirements.txt
kopf==1.36.0
kubernetes==26.1.0
PyYAML~=5.4.1